python-version/ImageProcessor/multi_image_layout/requirements.txt
PyQt5>=5.15.0
Pillow>=10.0.0
reportlab>=4.0.0
PyPDF2>=3.0.0
numpy>=1.24.0
